# packages/api-client/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FamilyPlaces API",
        "description": "Versioned public contract for the FamilyPlaces catalogue.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": ""
        }
    ],
    "paths": {
...
        "/api/v1/places/{slug}": {
            "get": {
                "operationId": "getPlaceBySlug",
                "tags": [
                    "Discovery"
                ],
                "responses": {
                    "200": {
                        "description": "Published place details.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": false,
                                    "required": [
                                        "id",
                                        "slug",
                                        "name",
                                        "short_description",
                                        "description",
                                        "city_name",
                                        "city_slug",
                                        "address_line1",
                                        "address_line2",
                                        "postal_code",
                                        "country_code",
                                        "categories",
                                        "amenities",
                                        "age_zones",
                                        "weekly_opening",
                                        "special_opening",
                                        "indoor",
                                        "outdoor",
                                        "free_entry",
                                        "price_description",
                                        "website_url",
                                        "phone",
                                        "verification_status",
                                        "longitude",
                                        "latitude"
                                    ],
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "slug": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "short_description": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "city_name": {
                                            "type": "string"
                                        },
                                        "city_slug": {
                                            "type": "string"
                                        },
                                        "address_line1": {
                                            "type": "string"
                                        },
                                        "address_line2": {
                                            "type": [
                                                "string",
                                                "null"
                                            ]
                                        },
                                        "postal_code": {
                                            "type": "string"
                                        },
                                        "country_code": {
                                            "type": "string"
                                        },
                                        "categories": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "additionalProperties": false,
                                                "required": [
                                                    "slug",
                                                    "name"
                                                ],
                                                "properties": {
                                                    "slug": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        },
                                        "amenities": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "additionalProperties": false,
                                                "required": [
                                                    "slug",
                                                    "name"
                                                ],
                                                "properties": {
                                                    "slug": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        },
                                        "age_zones": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "weekly_opening": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "special_opening": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "indoor": {
                                            "type": "boolean"
                                        },
                                        "outdoor": {
                                            "type": "boolean"
                                        },
                                        "free_entry": {
                                            "type": "boolean"
                                        },
                                        "price_description": {
                                            "type": [
                                                "string",
                                                "null"
                                            ]
                                        },
                                        "website_url": {
                                            "type": [
                                                "string",
                                                "null"
                                            ]
                                        },
                                        "phone": {
                                            "type": [
                                                "string",
                                                "null"
                                            ]
                                        },
                                        "verification_status": {
                                            "type": "string"
                                        },
                                        "longitude": {
                                            "type": "number"
                                        },
                                        "latitude": {
                                            "type": "number"
                                        },
                                        "main_photo": {
                                            "type": [
                                                "object",
                                                "null"
                                            ],
                                            "properties": {
                                                "thumbnail_mini": {
                                                    "type": "string"
                                                },
                                                "thumbnail": {
                                                    "type": "string"
                                                },
                                                "card": {
                                                    "type": "string"
                                                },
                                                "hero": {
                                                    "type": "string"
                                                },
                                                "original_max": {
                                                    "type": "string"
                                                }
                                            }
                                        },
                                        "photos": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid"
                                                    },
                                                    "is_main": {
                                                        "type": "boolean"
                                                    },
                                                    "alt_text": {
                                                        "type": [
                                                            "string",
                                                            "null"
                                                        ]
                                                    },
                                                    "caption": {
                                                        "type": [
                                                            "string",
                                                            "null"
                                                        ]
                                                    },
                                                    "variants": {
                                                        "type": "object",
                                                        "properties": {
                                                            "thumbnail_mini": {
                                                                "type": "string"
                                                            },
                                                            "thumbnail": {
                                                                "type": "string"
                                                            },
                                                            "card": {
                                                                "type": "string"
                                                            },
                                                            "hero": {
                                                                "type": "string"
                                                            },
                                                            "original_max": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "RFC 9457 place_not_found Problem Details.",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "type",
                                        "title",
                                        "status",
                                        "detail",
                                        "code",
                                        "correlationId"
                                    ],
                                    "properties": {
                                        "type": {
                                            "type": "string"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "integer"
                                        },
                                        "detail": {
                                            "type": "string"
                                        },
                                        "code": {
                                            "type": "string"
                                        },
                                        "correlationId": {
                                            "type": "string",
                                            "format": "uuid"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "summary": "Get published place details",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Published place slug.",
                        "required": true,
                        "deprecated": false,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple",
                        "explode": false
                    }
                ],
                "security": []
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {}
    },
    "security": [],
...
}